Update README.md
I used Piñata to store my index.html which included an embedded version of my Music NFT on OpenSea but it failed to display my NFT. It was done for experimental purposes only. A NFT within an NFT would be silly right? Still, there's zero reasons why experimentation should be limited to imagination. With that said I will keep this going, but not necessarily with the goal to store my NFT on IPFS as it already is. 

Conclusion: I thought inside the box, now let's go outside of it. This was done in a few minutes, so no loss and more knowledge gained!